+++ packages/musicos/sources/track_position.move
@@ -11,26 +11,26 @@ module musicos::track_position;
 /// Stores the disc index and track index as a tuple.
 public struct TrackPosition(
     /// Zero-based index of the disc within the release.
-    u8,
+    u64,
     /// Zero-based index of the track within the disc.
-    u8,
+    u64,
 ) has copy, drop, store;
 
 //=== Public Functions ===
 
 /// Creates a new track identifier with the specified disc and track indices.
-public fun new(disc_index: u8, track_index: u8): TrackPosition {
+public fun new(disc_index: u64, track_index: u64): TrackPosition {
     TrackPosition(disc_index, track_index)
 }
 
 //=== Public View Functions ===
 
 /// Returns the disc index component of this identifier.
-public fun disc_idx(self: TrackPosition): u8 {
+public fun disc_idx(self: TrackPosition): u64 {
     self.0
 }
 
 /// Returns the track index component of this identifier.
-public fun track_idx(self: TrackPosition): u8 {
+public fun track_idx(self: TrackPosition): u64 {
     self.1
 }
